Fixes #120.
* Folding `memory_device_impls.hpp` into `multi_wrapper_impls.hpp`
