+++ src/cuda/api_wrappers.hpp
@@ -30,7 +30,6 @@
 
 #include <cuda/api/peer_to_peer.hpp>
 
-#include <cuda/api/memory_device_impls.hpp>
 #include <cuda/api/pci_id_impl.hpp>
 #include <cuda/api/multi_wrapper_impls.hpp>
